Don't compile iconv on Windows
