+++ parse.go
@@ -1,17 +1,12 @@
 package mailaddress
 
 import (
-	"bytes"
 	"errors"
-	"io"
-	"io/ioutil"
 	"mime"
 	"regexp"
 	"strings"
 	"unicode"
 	"unicode/utf8"
-
-	iconv "gopkg.in/iconv.v1"
 )
 
 var (
@@ -146,31 +141,6 @@ func parse(str string) (list List, haveError bool) {
 	}
 
 	return list, haveError
-}
-
-// Convert the bytes from the input reader to UTF-8
-func toUTF8(charset string, input io.Reader) (_ io.Reader, returnErr error) {
-	conv, err := iconv.Open("utf-8", charset)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		if err := conv.Close(); err != nil {
-			returnErr = err
-		}
-	}()
-
-	r := iconv.NewReader(conv, input, 0)
-	b, err := ioutil.ReadAll(r)
-	if err != nil {
-		// errno 84 from syscall. Unfortunately we can't check the errno :-/
-		if err.Error() == "invalid or incomplete multibyte or wide character" {
-			return nil, ErrInvalidEncoding
-		}
-		return nil, err
-	}
-
-	return bytes.NewReader(b), returnErr
 }
 
 func end(a *Address) (goterror bool) {
